Add sequences for Aux and Output Sources - Aux Source - HDMI 1/ HDMI 2/ UVC Output - Add descriptions to most settings actions (unrelated to the sequences)
Note: these are primarily intended to be used a "simple" toggles in which the user selects a subset of the options only. For example, toggle Aux Source between UVC In and Player (skipping NDI), or toggle HDMI 1 between In 4 and PGM (something I actually do). While I can imagine a use-case for random cycling to HDMI out, such functionality is not currently implemented.